feat(wasm-utxo): restructure fixedScriptWallet into directory
Reorganize fixedScriptWallet into a proper directory structure with
separate files for distinct functionality. Move RootWalletKeys and
address functions into their own modules while keeping the same API
through the index file.

Issue: BTC-2786
