+++ packages/wasm-utxo/js/fixedScriptWallet/index.ts
@@ -1,0 +1,11 @@
+export { RootWalletKeys, type WalletKeysArg, type IWalletKeys } from "./RootWalletKeys.js";
+export { outputScript, address } from "./address.js";
+export {
+  BitGoPsbt,
+  type NetworkName,
+  type ScriptId,
+  type InputScriptType,
+  type ParsedInput,
+  type ParsedOutput,
+  type ParsedTransaction,
+} from "./BitGoPsbt.js";
