+++ packages/wasm-utxo/js/index.ts
@@ -7,7 +7,7 @@ void wasm;
 export * as address from "./address.js";
 export * as ast from "./ast/index.js";
 export * as utxolibCompat from "./utxolibCompat.js";
-export * as fixedScriptWallet from "./fixedScriptWallet.js";
+export * as fixedScriptWallet from "./fixedScriptWallet/index.js";
 export * as bip32 from "./bip32.js";
 export * as ecpair from "./ecpair.js";
